add mutation option functions
