+++ src/components/apollo/TodoItem.js
@@ -3,9 +3,10 @@ import { QuickForm, BoolField } from 'uniforms-antd';
 import { Card, Button } from 'antd';
 
 import { schema } from '../../schema';
-import { UPDATE_TODO, DELETE_TODO, FIND_TODOS } from '../../gql/queries';
+import { UPDATE_TODO, DELETE_TODO } from '../../gql/queries';
 import { OfflineContext } from "../../config/client.apollo";
 import { getOptimisticResponse } from '../../utils/CreateOptimisticResponse';
+import { updateTodoUpdateQuery, deleteTodoUpdateQuery } from '../../utils/mutationOptions';
 
 
 export function TodoItem({ todo }) {
@@ -18,22 +19,7 @@ export function TodoItem({ todo }) {
       query: UPDATE_TODO,
       variables,
       optimisticResponse: getOptimisticResponse(variables, UPDATE_TODO),
-      update: (proxy, { data: { updateTodo }}) => {
-        const data = proxy.readQuery({ query: FIND_TODOS });
-        const items = data.findTodos.items.map(item => {
-          if (item.id === updateTodo.id) return updateTodo;
-          return item;
-        });
-        proxy.writeQuery({ 
-          query: FIND_TODOS,
-          data: {
-            findTodos: {
-              ...data.findTodos,
-              items,
-            },
-          }
-        })
-      }
+      update: updateTodoUpdateQuery,
     }).subscribe(
       (res) => console.log('update sub', res),
       (err) => console.log('update error', err)
@@ -46,19 +32,7 @@ export function TodoItem({ todo }) {
       query: DELETE_TODO,
       variables,
       optimisticResponse: getOptimisticResponse(variables, DELETE_TODO),
-      update: (proxy, { data: { deleteTodo }}) => {
-        const data = proxy.readQuery({ query: FIND_TODOS });
-        const items = data.findTodos.items.filter(item => item.id !== deleteTodo.id);
-        proxy.writeQuery({ 
-          query: FIND_TODOS,
-          data: {
-            findTodos: {
-              ...data.findTodos,
-              items,
-            },
-          }
-        })
-      }
+      update: deleteTodoUpdateQuery,
     }).subscribe(
       (res) => console.log('delete sub', res),
       (err) => console.log('delete', err)
